// src/lib/sanity/types.ts
export type Exercise = {
  _id: string;
  _type: "exercise";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  description?: string;
  difficulty?: "beginner" | "intermediate" | "advanced";
  image?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    _type: "image";
  };
  videoUrl?: string;
  isActive?: boolean;
};
// ...
export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type SingleExerciseQueryResult = {
  _id: string;
  name: string | null;
  description: string | null;
  difficulty: "advanced" | "beginner" | "intermediate" | null;
  image: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    _type: "image";
  } | null;
  videoUrl: string | null;
  isActive: boolean | null;
} | null;